requirements.txt
requests==2.28.1
pillow==9.2.0
